src/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.10
python-dotenv==1.0.1
pydantic-settings==2.5.2
aiofile==3.8.8
langchain==0.1.20
PyMuPDF==1.24.10
motor==3.6.0
openai==1.51.2
cohere==5.11.0
qdrant-client==1.12.0
